refactor: ui 패키지 분리
