--- app/src/main/java/woowacourse/movie/MainActivity.kt
@@ -1,11 +1,0 @@
-package woowacourse.movie
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
