+++ app/src/main/java/woowacourse/movie/ui/fragment/movielist/adapter/MovieListViewType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.adapter
+package woowacourse.movie.ui.fragment.movielist.adapter
 
 enum class MovieListViewType {
     MOVIE_VIEW, AD_VIEW
