+++ app/src/main/res/drawable/bottom_item_settings.xml
