+++ app/src/main/res/layout/fragment_reservation_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.ReservationListFragment">
+    tools:context=".ui.fragment.reservationlist.ReservationListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_reservation"
