+++ app/src/main/res/menu/menu_main_bottom_navigation.xml
@@ -10,8 +10,8 @@
         android:icon="@drawable/bottom_item_home"
         android:title="@string/bottom_item_home" />
     <item
-        android:id="@+id/bottom_item_setting"
-        android:icon="@drawable/bottom_item_setting"
-        android:title="@string/bottom_item_setting" />
+        android:id="@+id/bottom_item_settings"
+        android:icon="@drawable/bottom_item_settings"
+        android:title="@string/bottom_item_settings" />
 
 </menu>
